# frontend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LineSight",
    "description": "LineSight: AI-driven Semantic ETL platform for SMB apparel manufacturing. Parse messy Excel files, track SAM/DHU metrics, and ensure UFLPA compliance.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dashboards/": {
      "post": {
        "tags": [
          "Dashboards"
        ],
        "summary": "Create Dashboard",
        "description": "Create a new dashboard with widget configuration.\n\n- **name**: Dashboard name (required)\n- **description**: Optional description\n- **data_source_id**: Optional link to DataSource\n- **widget_config**: Widget configuration (enabled widgets + settings)\n- **layout_config**: Grid layout configuration",
        "operationId": "create_dashboard_api_v1_dashboards__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Dashboards"
        ],
        "summary": "List Dashboards",
        "description": "List all dashboards for the current user.\nReturns dashboards ordered by most recently updated.\nOptional filtering by factory_id.",
        "operationId": "list_dashboards_api_v1_dashboards__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "factory_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter dashboards by factory",
              "title": "Factory Id"
            },
            "description": "Filter dashboards by factory"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/{dashboard_id}": {
      "get": {
        "tags": [
          "Dashboards"
        ],
        "summary": "Get Dashboard",
        "description": "Get a specific dashboard with widget data.\n\nReturns the dashboard configuration plus actual widget data\nfetched from the linked data source, including production line ID for filtering.",
        "operationId": "get_dashboard_api_v1_dashboards__dashboard_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dashboards"
        ],
        "summary": "Update Dashboard",
        "description": "Update dashboard configuration.\n\nAllows updating name, description, data source, widget config, and layout.\nOnly provided fields will be updated.",
        "operationId": "update_dashboard_api_v1_dashboards__dashboard_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Dashboards"
        ],
        "summary": "Delete Dashboard",
        "description": "Delete a dashboard.\n\nThis will remove the dashboard but not the linked data source.",
        "operationId": "delete_dashboard_api_v1_dashboards__dashboard_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "description": "Dashboard name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Dashboard description"
          },
          "data_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Source Id",
            "description": "UUID of the linked ExcelUpload"
          },
          "widget_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WidgetConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Widget configuration"
          },
          "layout_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LayoutConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Layout configuration"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DashboardCreate",
        "description": "Schema for creating a new dashboard."
      },
      "DashboardDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "data_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Source Id"
          },
          "widget_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Widget Config"
          },
          "layout_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout Config"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "widget_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Widget Data",
            "description": "Actual widget data fetched from data source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "DashboardDetailResponse",
        "description": "Detailed dashboard response with widget data."
      },
      "DashboardListResponse": {
        "properties": {
          "dashboards": {
            "items": {
              "$ref": "#/components/schemas/DashboardResponse"
            },
            "type": "array",
            "title": "Dashboards"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "dashboards",
          "count"
        ],
        "title": "DashboardListResponse",
        "description": "List of dashboards for a user."
      },
      "DashboardResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "data_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Source Id"
          },
          "widget_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Widget Config"
          },
          "layout_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout Config"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "DashboardResponse",
        "description": "Basic dashboard response schema."
      },
      "DashboardUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "data_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Source Id"
          },
          "widget_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WidgetConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "layout_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LayoutConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "DashboardUpdate",
        "description": "Schema for updating an existing dashboard."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LayoutConfig": {
        "properties": {
          "layouts": {
            "items": {
              "$ref": "#/components/schemas/LayoutItem"
            },
            "type": "array",
            "title": "Layouts",
            "description": "Widget layout items"
          }
        },
        "type": "object",
        "title": "LayoutConfig",
        "description": "Dashboard layout configuration."
      },
      "LayoutItem": {
        "properties": {
          "widget_id": {
            "type": "string",
            "title": "Widget Id",
            "description": "Unique identifier for the widget"
          },
          "x": {
            "type": "integer",
            "minimum": 0.0,
            "title": "X",
            "description": "X position in grid"
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y",
            "description": "Y position in grid"
          },
          "w": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "W",
            "description": "Width in grid units"
          },
          "h": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "H",
            "description": "Height in grid units"
          }
        },
        "type": "object",
        "required": [
          "widget_id",
          "x",
          "y",
          "w",
          "h"
        ],
        "title": "LayoutItem",
        "description": "Widget layout position and size in grid."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WidgetConfig": {
        "properties": {
          "enabled_widgets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enabled Widgets",
            "description": "List of enabled widget types"
          },
          "widget_settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Widget Settings",
            "description": "Per-widget custom settings"
          }
        },
        "type": "object",
        "title": "WidgetConfig",
        "description": "Dashboard widget configuration."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}